requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
requests==2.31.0
cryptography==41.0.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
rsa==4.9
six==1.16.0
urllib3==2.0.7